refactor: move WorldPathing.py to adapter dir, update WorldMap+ import via sys.path
